Update namelist.input
Minor updates to the Hurricane Matthew test case WRF namelist, none of which will cause WRF to run any differently. The changes reduce potential confusion for users (only the first column is being read in this 1-domain simulation, but the dates in the second column were made the same as the first for clarity), and add commas to the end of the lines in which they were previously omitted (in case the absence of a comma confuses some text parsers).